Remove race condition
Update tasks so that they accept PK's as arguments, that way we don't
have to use pickles in celery any more

Update tasks so that a parent repository will start a group of
subtasks to create its own forks. This removes the need for a call to
AsyncResult.wait() within a celery task (this will be impossible to
call from tasks soon anyway).

Update post_save signal receiver to create repos using the new tasks